+++ app/src/main/java/com/kyoapps/maniac/helpers/classes/LoadRequestItem.kt
@@ -1,7 +1,8 @@
 package com.kyoapps.maniac.helpers.classes
 
+//@JsonClass(generateAdapter = true)
 data class LoadRequestItem(
-    val brdid: Short,
+    val brdid: Int,
     val thrdid: Int?,
     val msgid: Int?
 )
