+++ app/src/main/java/com/kyoapps/maniac/helpers/classes/pojo/Board.kt
@@ -3,6 +3,6 @@ package com.kyoapps.maniac.helpers.classes.pojo
 
 data class Board(
 
-        val brdid: Short,
+        val brdid: Int,
         val label: String
 )
